+++ ricaun.RevitTest.Command/Process/RevitTestProcessStartUtils.cs
@@ -52,7 +52,7 @@ namespace ricaun.RevitTest.Command.Process
             string applicationPath,
             string file,
             Action<TestModel> testResultAction,
-            int version = 0,
+            string version = null,
             string language = null,
             bool revitOpen = false,
             bool revitClose = false,