+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] / 2025-02-07 - 2025-02-19
+### Features
+- Update `ricaun.NUnit` support `TestFixture` - Version 1.5.0.
+- Update `ricaun.Revit.UI` version 0.8.0.
+### Commands
+- Update `RevitTestProcessStart` to use `nameof`.
+- Update `Options.RevitVersion` to be string. (Fix: #68)
+### Console
+- Update `RevitTestUtils` version to use string. (Fix: #68)
+### TestAdapter
+- Update `TestCaseUtils.SplitTestName` to split and ignore `(` and `)`.
+- Update `NUnit.Version` to be string. (Fix: #68)
+### Tests
+- Add `TestsFixture` and `TestFixtureSource`.
+- Add `TestsRevitFixtureSource` with Revit parameter and `TestFixtureSource` attribute.
+- Add `TestsFixtureParameterized` with multiple constructors.
+
 ## [1.8.0] / 2024-12-16 - 2024-12-19
 ### Features
 - Make project public
@@ -569,6 +586,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [x] TestsFail
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.9.0]: ../../compare/1.8.0...1.9.0
 [1.8.0]: ../../compare/1.7.1...1.8.0
 [1.7.1]: ../../compare/1.7.0...1.7.1
 [1.7.0]: ../../compare/1.6.0...1.7.0